api/requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Deprecated==1.2.13
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
kombu==5.2.3
lxml==4.7.1
packaging==21.3
parsel==1.6.0
priority==1.3.0
prompt-toolkit==3.0.27
Protego==0.1.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
queuelib==1.6.2
redis==4.1.2
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
typing-extensions==4.0.1
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
